+++ lib/review/ui/screens/book_search_screen.dart
@@ -6,8 +6,8 @@ import 'package:typed/common/index.dart';
 import 'package:typed/review/ui/components/custom_placeholder.dart';
 import 'package:typed/review/ui/components/custom_progress_indicator.dart';
 import 'package:typed/review/ui/screens/review_error_screen.dart';
+import 'package:typed/review/ui/widgets/book_search_result_widget.dart';
 import 'package:typed/review/viewmodels/book_providers.dart';
-import 'package:typed/review/data/models/book_model.dart';
 
 class BookSearchScreen extends ConsumerStatefulWidget {
   const BookSearchScreen({super.key});
@@ -39,13 +39,6 @@ class _BookSearchScreenState extends ConsumerState<BookSearchScreen> {
     _searchController.dispose();
     _searchFocusNode.dispose();
     super.dispose();
-  }
-
-  void _performSearch() {
-    final query = _searchController.text.trim();
-    if (query.isNotEmpty) {
-      ref.read(searchQueryProvider.notifier).state = query;
-    }
   }
 
   @override
@@ -86,7 +79,13 @@ class _BookSearchScreenState extends ConsumerState<BookSearchScreen> {
                   itemCount: books.length,
                   itemBuilder: (context, index) {
                     final book = books[index];
-                    return _buildBookItem(context, book);
+                    return BookSearchResultWidget(
+                      book: book,
+                      onTap: () {
+                        ref.read(selectedBookProvider.notifier).state = book;
+                        context.push('/review_input');
+                      },
+                    );
                   },
                 );
               },
@@ -99,71 +98,6 @@ class _BookSearchScreenState extends ConsumerState<BookSearchScreen> {
           ),
         ],
       ),
-    );
-  }
-
-  Widget _buildBookItem(BuildContext context, Book book) {
-    return GestureDetector(
-      onTap: () {
-        ref.read(selectedBookProvider.notifier).state = book;
-        context.push('/review_input');
-      },
-      child: Container(
-        width: double.infinity,
-        padding: const EdgeInsets.symmetric(horizontal: 16.0, vertical: 8.0),
-        child: Row(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Container(
-              width: 80,
-              height: 120,
-              decoration: BoxDecoration(
-                color: AppColors.backgroundTertiary,
-                borderRadius: BorderRadius.zero,
-                border: Border.all(
-                  color: AppColors.borderBlack,
-                  width: 0.3,
-                ),
-              ),
-              child: book.thumbnail.isNotEmpty
-                  ? Image.network(
-                      book.thumbnail,
-                      fit: BoxFit.cover,
-                      errorBuilder: (context, error, stackTrace) {
-                        return Center(
-                          child: CustomPlaceholder(size: 0.06),
-                        );
-                      },
-                    )
-                  : Center(
-                      child: CustomPlaceholder(size: 0.1),
-                    ),
-            ),
-            const SizedBox(width: 16),
-            Expanded(
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  _buildBookItemText(book.title),
-                  const SizedBox(height: 4),
-                  _buildBookItemText(book.author),
-                  const SizedBox(height: 4),
-                  _buildBookItemText(book.publisher),
-                ],
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Widget _buildBookItemText(String text) {
-    return Text(
-      text,
-      style: AppTheme.body2,
-      maxLines: 1,
-      overflow: TextOverflow.ellipsis,
     );
   }
 
@@ -231,5 +165,12 @@ class _BookSearchScreenState extends ConsumerState<BookSearchScreen> {
         ],
       ),
     );
+  }
+
+  void _performSearch() {
+    final query = _searchController.text.trim();
+    if (query.isNotEmpty) {
+      ref.read(searchQueryProvider.notifier).state = query;
+    }
   }
 }
